add overall debug flag to indicate debugging level
currently there is no way to adjust the debug level of the cli
without recompilation of the cli.

this change adds a debug flag '--debug' or '-d' that allows the
verbosity of the output to be set.

by default, all debug messages should be written to stderr so that
stdout can be piped to the 'oc' tool if needed.

minor changes were also made to not use the smtpdetails errors in
the sendgrid api logic. this will allow sendgrid api logic to be
separated into a separate package in the future. resulting in a
package structure of:

'''
pkg/
  sendgrid/     - for rhmi-aware sendgrid logic
  sendgridapi/  - for independent sendgrid logic
  smtpdetails/  - for abstracted interfaces and structs
'''

this restructure does not take place in this commit.

verification:
- run './cli create rhmi_deleteme --debug=verbose'
- ensure debug logs are shown and are written to stderr
- run './cli create rhmi_deleteme'
- ensure there are no debug logs by default