add secret kind and version explicitly when converting
currently the kind and version in the type meta of the secret
output is not set. this means that the output of the 'create'
command is not a valid secret and cannot be piped into the
'oc create -f -' command. this is something the cli should be
able to do.

this commit updates the secret to ensure the kind and version are
set in the type metadata on the secret resource.

verification:
- authenticate with an openshift cluster using 'oc login ...'
- run './cli delete rhmi_deleteme' to cleanup an existing user
- run './cli create rhmi_deleteme | oc create -f -'
- run 'oc get secret rhmi-smtp'
- ensure a new secret is created in the openshift cluster