add goreleaser to push binaries on each tag
currently, to make binaries for the smtp-service cli available
they must be manually built. this will eventually lead to human
error and inconsistencies in the release format.

this commit adds a github workflow that will be performed on each
tag of the repo that will run the goreleaser tool. this will build
and push binaries to github releases automatically.

verification:
- create a new tag
- ensure the release workflow starts on github
- ensure new binaries are pushed for the new tag