add prerequisite check for sendgrid api key on default client
when a default sendgrid wrapper client is created, the env var
SENDGRID_API_KEY is used as the value of the api key passed into
the client.

it is possible that this env var is not set.

this commit ensures a meaningful error message is returned when
the SENDGRID_API_KEY env var is not set.

verification:
- run 'make build/cli'
- run 'unset SENDGRID_API_KEY'
- run './cli'
- ensure the error printed out explains the issue correctly