fix version command needing sendgrid credentials to be set in cli
currently the cli requires the sendgrid api key to be set before
even the 'version' command could be run. this key is not needed
in the case of some commands, such as 'version'.

this fixes this error by only setting up smtp details client in
commands where it is required.

verification:
- run 'make build/cli'
- run 'unset SENDGRID_API_KEY'
- run './cli version'
- ensure the version command completes successfully